+++ Rhino.DivanDB.Tests/Storage/CreateIndexes.cs
@@ -19,14 +19,14 @@ namespace Rhino.DivanDB.Tests.Storage
                 @"
     from doc in docs
     where doc.type == ""page""
-    select new { Key = doc.title, Value = doc.content, Size = (int)doc.size };
+    select new { Key = doc.title, Value = doc.content, Size = doc.size };
 ");
 
             db.PutIndex("pagesByTitle",
                 @"
     from doc in docs
     where doc.type == ""page""
-    select new { Key = doc.title, Value = doc.content, Size = (int)doc.size };
+    select new { Key = doc.title, Value = doc.content, Size = doc.size };
 ");
         }
 
@@ -37,7 +37,7 @@ namespace Rhino.DivanDB.Tests.Storage
                 @"
     from doc in docs
     where doc.type == ""page""
-    select new { Key = doc.title, Value = doc.content, Size = (int)doc.size };
+    select new { Key = doc.title, Value = doc.content, Size = doc.size };
 ");
             var indexNames = db.IndexDefinitionStorage.IndexNames;
             Assert.Equal(1, indexNames.Length);
@@ -50,7 +50,7 @@ namespace Rhino.DivanDB.Tests.Storage
             const string definition = @" 
     from doc in docs
     where doc.type == ""page""
-    select new { Key = doc.title, Value = doc.content, Size = (int)doc.size };
+    select new { Key = doc.title, Value = doc.content, Size = doc.size };
 ";
             db.PutIndex("pagesByTitle", definition);
             var actualDefinition = db.IndexDefinitionStorage.GetIndexDefinition("pagesByTitle");
